Put common knapsack initializers for lns cp and lp in a "generic" lns module
